Refactoring
Found a bug in the refactoring renaming function:
Renamed all the names in the file including the name of the
file in the explorer BUT NOT THE NAME OF THE FILE IN THE FINDER.
File would then not open. Fixed by going into finder and manually
rnaming the file directly.